Fixed broken pointer handling that lead to crashing.
When the snake escaped, the code tried to free the memory of the global variables in the main code. This error was made because of confusion regarding the fact that while the structs were created on the heap, dereferncing the result of the "constructor" only copied this value to the global variables. Thus, freeing them made no sense. Also, this fixes a minor memory leak. ;)